06_flask_fastapi/queimadas_flask/pyproject.toml
[project]
name = "queimadas-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-sql-connector[pyarrow]>=4.2.5",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
]
